ROS: add example and fix infrastructure.
This commit adds ROS 2.0 code as an additional
module. The code gets compiled automatically by
the CMake system and the corresponding binaries
are generated however the final px4 binary fails
to be generated because the program looks for a
library. Such library is currently not generated
from the ROS 2.0 code.